+++ app/src/main/kotlin/com/stevenfrew/beatprompter/ui/SongListAdapter.kt
@@ -2,9 +2,9 @@ package com.stevenfrew.beatprompter.ui
 
 import android.content.Context
 import com.stevenfrew.beatprompter.BeatPrompter
-import com.stevenfrew.beatprompter.graphics.bitmaps.AndroidBitmap
 import com.stevenfrew.beatprompter.graphics.bitmaps.Bitmap
 import com.stevenfrew.beatprompter.set.PlaylistNode
+import com.stevenfrew.beatprompter.util.Utils
 
 class SongListAdapter(
 	values: List<PlaylistNode>,
@@ -19,6 +19,8 @@ class SongListAdapter(
 	override val songIconDisplayPosition = BeatPrompter.preferences.songIconDisplayPosition
 	override val showMusicIcon = BeatPrompter.preferences.showMusicIcon
 
-	override fun getIconBitmap(icon: String): android.graphics.Bitmap =
-		(imageDictionary[icon] as AndroidBitmap?)?.androidBitmap ?: missingIconBitmap
+	override fun getIconBitmap(
+		icon: String?,
+		tags: Set<String>,
+	) = Utils.getIconBitmap(icon, tags, imageDictionary, missingIconBitmap)
 }